ui-comparison-test/requirements.txt
pytest>=7.4.0
playwright>=1.40.0
pytest-asyncio>=0.21.0
Pillow>=10.0.0
numpy>=1.24.0
PyYAML>=6.0

